==> src/scripts/05_date_dimension.ipynb <==
from datetime import datetime
import pandas as pd

start_date = datetime(2023, 1, 1)
end_date = datetime(2028, 1, 1)

date_range = pd.date_range(start_date, end_date)
date_df = pd.DataFrame(date_range, columns=['Date'])

# how to get year / month / day from the date
date_df['year'] = date_df['Date'].dt.year
date_df['month'] = date_df['Date'].dt.month
date_df['day'] = date_df['Date'].dt.day
date_df['day_of_week'] = date_df['Date'].dt.dayofweek + 1     # +1 to start from 1 
date_df['is_week_end'] = date_df['day_of_week'].isin([6, 7])  # gives True for Saturday and Sunfday



#date_df
# index column would cause problems in AWS, so all the time skip
# the writing of the index to the csv file
date_df.to_csv(r'..\..\csv\date_dimension.csv', index=False)
