==> src/scripts/06_chicago_data_to_mapping.ipynb <==
# some exploratory data analysis
from datetime import datetime
from dateutil.relativedelta import relativedelta
import os

import pandas as pd
# set pandas to show 100 the columns
pd.set_option('display.max_columns', 100)

import requests


# get the taxi data of the last full month (T-1 months') data
current_datetime = datetime.now()    # current date
# get the data two months before as a formatted string
formatted_datetime = (current_datetime - relativedelta(months=2)).strftime('%Y-%m-%d')

url = f'https://data.cityofchicago.org/resource/wrvz-psew.json?$where=trip_start_timestamp >= "{formatted_datetime}T00:00:00" AND trip_start_timestamp <= "{formatted_datetime}T23:59:59"&$limit=30000'

# the os.environ.get looks for the specified variable in the .env file of the root
# folder of the project
headers = {'X-App-Token': os.environ.get("CHICHAGO_API_TOKEN")}

# response = requests.get(url, headers)    # in case an error comes leave the headers parameter
response = requests.get(url)

data = response.json()

#type(response)    # we can check the type: "requests.models.Response"

# create a dataframe from the taxi trip data
taxi_trips = pd.DataFrame(data)
taxi_trips.head()

taxi_trips.info()

taxi_trips.describe()

# check rows where the fare is null, look at just 5 random samples
taxi_trips[taxi_trips['fare'].isna()].sample(5)

# in this case ask for advice from the potential users what to do

# some conclusions:
# - total trip might be erronous
# - fare data are missing from a few raws
# - the payment type and company occupies too much space
# - the centroid locations as Point coordinates are duplicates
# - the census tract data are very sparse, more than half of them are missing


#### Transformation: deal with the NaN values

# the order is important, because these two columns contain lots of NaN values
# first drop them, then the rows having NaN in the other columns only
taxi_trips.drop(['pickup_census_tract', 'dropoff_census_tract'], axis=1, inplace=True)
taxi_trips.dropna(inplace=True)

#### Transformation: drop the unnecessary duplicated columns

taxi_trips.drop(['pickup_centroid_location', 'dropoff_centroid_location'], axis=1, inplace=True)

#### Transformation: renaming

taxi_trips.rename(columns={"pickup_community_area": "pickup_community_area_id",
                           "dropoff_community_area": "dropoff_community_area_id"},
                           inplace=True)

Transformation: an auxilary column to the weather data

taxi_trips['trip_start_timestamp'] = pd.to_datetime(taxi_trips['trip_start_timestamp'])
taxi_trips['datetime_for_weather'] = taxi_trips['trip_start_timestamp'].dt.floor('H')

#### Check joining the trips and the weather data

# get weather data from the open-meteo.com site, and create a dataframe out if it
url = 'https://archive-api.open-meteo.com/v1/era5'
# current date
current_datetime = datetime.now()
# get the data two months before as a formatted string
formatted_datetime = (current_datetime - relativedelta(months=2)).strftime('%Y-%m-%d')
params = {
    'latitude': 41.85,
    'longitude': -87.65,
    'start_date': formatted_datetime,
    'end_date': formatted_datetime,
    'hourly': 'temperature_2m,wind_speed_10m,rain,precipitation'
    }

response = requests.get(url, params=params)
weather_data = response.json()

# transform the JSON string to a pd dataframe
weather_data_filtered = {
    'datetime': weather_data['hourly']['time'],
    'temperature': weather_data['hourly']['temperature_2m'],
    'wind_speed': weather_data['hourly']['wind_speed_10m'],
    'rain': weather_data['hourly']['rain'],
    'precipitation': weather_data['hourly']['precipitation']
    }

#weather_data_filtered
weather_df = pd.DataFrame(weather_data_filtered)
weather_df['datetime'] = pd.to_datetime(weather_df['datetime'])
weather_df.info()
weather_df.head()


taxi_trips_with_weather = taxi_trips.merge(weather_df, left_on='datetime_for_weather', right_on='datetime')


#taxi_trips_with_weather.head()

Transformation: Data type conversions

# we can characterize columns which are not object is the dataframe
taxi_trips.describe()

data_types = {
    'trip_end_timestamp': 'datetime64[ns]',
    'trip_seconds': 'int32',
    'trip_miles': 'float',
    'pickup_community_area_id': 'int8',
    'dropoff_community_area_id': 'int8',
    'fare': 'float',
    'tips': 'float',
    'tolls': 'float',
    'extras': 'float',
    'trip_total': 'float',
}
taxi_trips = taxi_trips.astype(data_types)

taxi_trips.info()

# much better, the non-object columns have some basic statistics easily
taxi_trips.describe()
# calculate the memory usage of the dataframe
taxi_trips.memory_usage(deep=True).sum()
# the data type optimized memory usage is less by a 40%

#### Sanity checks

taxi_trips.describe()

taxi_trips[taxi_trips['trip_end_timestamp'] == taxi_trips['trip_end_timestamp'].max()]

taxi_trips[taxi_trips['trip_seconds'] == taxi_trips['trip_seconds'].max()]

taxi_trips[taxi_trips['fare'] == taxi_trips['fare'].max()]

taxi_trips.nlargest(10, 'fare')

#### Data modelling
Create master tables for the payment types and the taxi company names

# master table for the payment types
taxi_trips['payment_type'].unique()
payment_type_master = taxi_trips['payment_type'].drop_duplicates().reset_index(drop=True)


payment_type_master = pd.DataFrame(
    {
        'payment_type_id': range(1, len(payment_type_master)+1),    # the id's usually starts from 1 despite indeces starting from 0
        'payment_type': payment_type_master
    }
)

payment_type_master

# master table for the company names
taxi_trips['company'].unique()
company_master = taxi_trips['company'].drop_duplicates().reset_index(drop=True)


company_master = pd.DataFrame(
    {
        'company_id': range(1, len(company_master)+1),    # the id's usually starts from 1 despite indeces starting from 0
        'company': company_master
    }
)

company_master

Check the memory usage if we replace the long strings with small integers

#### Check the modelling
##### Payment type

print(taxi_trips.memory_usage(deep=True).sum())

taxi_trips_id = taxi_trips.merge(payment_type_master, on='payment_type')
taxi_trips_id = taxi_trips_id.merge(company_master, on='company')
taxi_trips_id.drop(['payment_type', 'company'], axis=1, inplace=True)
print(taxi_trips_id.memory_usage(deep=True).sum())

taxi_trips_id.sample(5)

# write the master tables to csv files
payment_type_master.to_csv(r'..\..\csv\payment_type_master.csv', index=False)
company_master.to_csv(r'..\..\csv\company_master.csv', index=False)

# master table for the payment types
taxi_trips['payment_type'].unique()
payment_type_master = taxi_trips['payment_type'].drop_duplicates().reset_index(drop=True)


payment_type_master = pd.DataFrame(
    {
        'payment_type_id': range(1, len(payment_type_master)+1),    # the id's usually starts from 1 despite indeces starting from 0
        'payment_type': payment_type_master
    }
)

payment_type_master

# how to extend the payment_type_master table if a new type of payment appears?
new_payment_type_data = [
    {'payment_type': 'Credit Card'},
    {'payment_type': 'X'},
    {'payment_type': 'Y'}
]
new_payment_type_mapping = pd.DataFrame(new_payment_type_data)

# payment_type_max_id = payment_type_master['payment_type_id'].max()

new_payment_type_mapping['payment_type_id'] =\
    range(payment_type_max_id + 1,
          payment_type_max_id + len(new_payment_type_mapping) + 1)

new_payment_type_mapping['payment_type'].isin(payment_type_master['payment_type'])

# filter only the payment types which are missing from the payment type master
# by default the filtering select the rows for which the evaluation is true, but
# using ~ we negate the results, so this line filters out from the actual list of
# payment types which are missing from our master table
new_payment_type_mapping[~new_payment_type_mapping['payment_type'].isin(payment_type_master['payment_type'])]
new_payment_types = new_payment_type_mapping[~new_payment_type_mapping['payment_type'].isin(payment_type_master['payment_type'])]

if not new_payment_types.empty:
    payment_type_master = pd.concat([payment_type_master, new_payment_types], ignore_index=True)
payment_type_master

#### Check the modelling
##### Company

# master table for the company names
taxi_trips['company'].unique()
company_master = taxi_trips['company'].drop_duplicates().reset_index(drop=True)


company_master = pd.DataFrame(
    {
        'company_id': range(1, len(company_master)+1),    # the id's usually starts from 1 despite indeces starting from 0
        'company': company_master
    }
)

company_master

# how to extend the company_master table if a new type of payment appears?
new_company_data = [
    {'company': 'Petani Cab Corp'},
    {'company': 'X'},
    {'company': 'Y'}
]
new_company_mapping = pd.DataFrame(new_company_data)

company_max_id = company_master['company_id'].max()

new_company_mapping['company_id'] =\
    range(company_max_id + 1,
          company_max_id + len(new_company_mapping) + 1)

new_company_mapping

new_company_mapping['company'].isin(company_master['company'])

# filter only the companies which are missing from the company master
# by default the filtering select the rows for which the evaluation is true, but
# using ~ we negate the results, so this line filters out from the actual list of
# companies which are missing from our master table
new_company_mapping[~new_company_mapping['company'].isin(company_master['company'])]
new_companies = new_company_mapping[~new_company_mapping['company'].isin(company_master['company'])]

if not new_companies.empty:
    company_master = pd.concat([company_master, new_companies], ignore_index=True)
company_master
