==> src/scripts/07_json_transform_load.ipynb <==
# some exploratory data analysis
from datetime import datetime
from dateutil.relativedelta import relativedelta
import json
import os

import pandas as pd
# set pandas to show 100 the columns
pd.set_option('display.max_columns', 100)

import requests


"""
1. get the data from S3
2. transform the weather data (weather_data)
3. transform the taxi data (taxi_trips) - DONE
4. update payment type master table (payment_type_master) - DONE
5. update company master table (company_master) - DONE
6. update the ids in the taxi_trips using the latest master tables - DONE
7. upload weather data fo S3 (weather_data)
8. upload taxi data fo S3 (taxi_trips)
9. upload the new payment and company master tables (payment_type_master, company_master)
"""

#### 3. transform the taxi data (taxi_trips)

# get the taxi data of the last full month (T-1 months') data
current_datetime = datetime.now()    # current date
# get the data two months before as a formatted string
formatted_datetime = (current_datetime - relativedelta(months=2)).strftime('%Y-%m-%d')

url = f'https://data.cityofchicago.org/resource/wrvz-psew.json?$where=trip_start_timestamp >= "{formatted_datetime}T00:00:00" AND trip_start_timestamp <= "{formatted_datetime}T23:59:59"&$limit=30000'

# the os.environ.get looks for the specified variable in the .env file of the root
# folder of the project
headers = {'X-App-Token': os.environ.get("CHICHAGO_API_TOKEN")}

# response = requests.get(url, headers)    # in case an error comes leave the headers parameter
response = requests.get(url)

data = response.json()


# create a dataframe from the taxi trip data
taxi_trips = pd.DataFrame(data)

taxi_trips.columns

# # the order is important, because these two columns contain lots of NaN values
# # first drop them, then the rows having NaN in the other columns only
# taxi_trips.drop(['pickup_census_tract', 'dropoff_census_tract'], axis=1, inplace=True)
# taxi_trips.drop(['pickup_centroid_location', 'dropoff_centroid_location'], axis=1, inplace=True)
# taxi_trips.dropna(inplace=True)
# taxi_trips.rename(columns={"pickup_community_area": "pickup_community_area_id",
#                            "dropoff_community_area": "dropoff_community_area_id"},
#                            inplace=True)
taxi_trips['trip_start_timestamp'] = pd.to_datetime(taxi_trips['trip_start_timestamp']).dt.floor('H')
taxi_trips['datetime_for_weather'] = taxi_trips['trip_start_timestamp'].dt.floor('H')

#### taxi_trips transformations function

def taxi_trips_transformations(taxi_trips: pd.DataFrame) -> pd.DataFrame:
    """ Perform transformation with the taxi data

    Args:
        taxi_trips (pd.DataFrame): 
            The DataFrame holding the daily taxi trips.

    Returns:
        pd.DataFrame:
            The cleaned, DataFrame holding the daily taxi trips.
    """
    
    # error handling - can be extended with some other ideas
    if not isinstance(taxi_trips, pd.DataFrame):
        raise TypeError('taxi_trips is a not a valid Pandas DataFrame.')
    
    # the order is important, because these two columns contain lots of NaN values
    # first drop them, then the rows having NaN in the other columns only
    taxi_trips.drop(['pickup_census_tract','dropoff_census_tract',
                     'pickup_centroid_location', 'dropoff_centroid_location'],
                    axis=1, inplace=True)
    taxi_trips.dropna(inplace=True)
    taxi_trips.rename(columns={"pickup_community_area": "pickup_community_area_id",
                            "dropoff_community_area": "dropoff_community_area_id"},
                            inplace=True)
    taxi_trips['trip_start_timestamp'] = pd.to_datetime(taxi_trips['trip_start_timestamp']).dt.floor('H')
    
    return taxi_trips

taxi_trips_transformations(taxi_trips)

#### company update codes

company_master = taxi_trips['company'].drop_duplicates().reset_index(drop=True)


company_master = pd.DataFrame(
    {
        'company_id': range(1, len(company_master)+1),    # the id's usually starts from 1 despite indeces starting from 0
        'company': company_master
    }
)

company_master

# how to extend the company_master table if a new type of payment appears?
new_company_data = [
    {'company': 'Petani Cab Corp'},
    {'company': 'X'},
    {'company': 'Y'}
]
new_company_mapping = pd.DataFrame(new_company_data)

company_max_id = company_master['company_id'].max()

# get the list of new companies
list_of_new_companies = []

for company in new_company_mapping['company'].values:   # go through the list of the latest list of companies
    if company not in company_master['company'].values: # find if any is missing from our master table
        list_of_new_companies.append(company)           # add it to the list of new companies

# the same in one line
list_of_new_companies_one_line = [company for company in new_company_mapping['company'].values
                                  if company not in company_master['company'].values]
list_of_new_companies_one_line

new_companies_df = pd.DataFrame({
    'company_id': range(company_max_id + 1, company_max_id + len(list_of_new_companies_one_line) + 1),
    'company': list_of_new_companies_one_line
})
new_companies_df

updated_company_master = pd.concat([company_master, new_companies_df], ignore_index=True)
updated_company_master.tail()

def update_company_master(taxi_trips: pd.DataFrame, company_master: pd.DataFrame) -> pd.DataFrame :
    """ Extend the company master with new companies if there are any of them.

    Args:
        taxi_trips (pd.DataFrame):
            DataFrame holding the daily taxi trips.
        company_master (pd.DataFrame):
            DataFrame holding the company_master data.

    Returns:
        pd.DataFrame:
            The updated company master data. If no new company appeared returns the original one.
    """
    company_max_id = company_master['company_id'].max()
    list_of_new_companies_one_line = [company for company in taxi_trips['company'].values
                                  if company not in company_master['company'].values]
    new_companies_df = pd.DataFrame({
        'company_id': range(company_max_id + 1, company_max_id + len(list_of_new_companies_one_line) + 1),
        'company': list_of_new_companies_one_line
        })
    updated_company_master = pd.concat([company_master, new_companies_df], ignore_index=True)
    
    return updated_company_master

taxi_trips_company_only = pd.DataFrame({
    'company_id': range(1,4),
    'company':     ['Petani Cab Corp', 'X', 'Y'] 
})
taxi_trips_company_only

updated_company_master = update_company_master(taxi_trips=taxi_trips_company_only, company_master=company_master)
updated_company_master

#### payment type update codes

payment_type_master = taxi_trips['payment_type'].drop_duplicates().reset_index(drop=True)


payment_type_master = pd.DataFrame(
    {
        'payment_type_id': range(1, len(payment_type_master)+1),    # the id's usually starts from 1 despite indeces starting from 0
        'payment_type': payment_type_master
    }
)


taxi_trips_payment_type_only = pd.DataFrame({
    'payment_type_id': range(1,4),
    'payment_type':     ['Credit Card', 'X', 'Y'] 
})
payment_type_master
taxi_trips_payment_type_only

def update_payment_type_master(taxi_trips: pd.DataFrame, payment_type_master: pd.DataFrame) -> pd.DataFrame :
    """ Extend the payment_type master with new payment types if there are any of them.

    Args:
        taxi_trips (pd.DataFrame):
            DataFrame holding the daily taxi trips.
        payment_type_master (pd.DataFrame):
            DataFrame holding the payment_type_master data.

    Returns:
        pd.DataFrame:
            The updated payment_type master data. If no new payment type appeared returns the original one.
    """
    payment_type_max_id = payment_type_master['payment_type_id'].max()
    list_of_new_payment_types_one_line = [payment_type for payment_type in taxi_trips['payment_type'].values
                                  if payment_type not in payment_type_master['payment_type'].values]
    new_payment_types_df = pd.DataFrame({
        'payment_type_id': range(payment_type_max_id + 1, payment_type_max_id + len(list_of_new_payment_types_one_line) + 1),
        'payment_type': list_of_new_payment_types_one_line
        })
    updated_payment_type_master = pd.concat([payment_type_master, new_payment_types_df], ignore_index=True)
    
    return updated_payment_type_master

taxi_trips['payment_type']

updated_payment_type_master = update_payment_type_master(taxi_trips=taxi_trips_payment_type_only, payment_type_master=payment_type_master)
updated_payment_type_master

#### combine the update functions to a more generic one

def update_master(taxi_trips: pd.DataFrame, master: pd.DataFrame, id_column: str,
                  value_column: str) -> pd.DataFrame :
    """ Extend the master with new types if there are any of them.

    Args:
        taxi_trips (pd.DataFrame):
            DataFrame holding the daily taxi trips.
        master (pd.DataFrame):
            DataFrame holding the master data.
        id_column (str):
            The id column of the master DataFrame.
        value_column (str): 
            The value column of the master and taxi_trips DataFrame.

    Returns:
        pd.DataFrame:
            The updated master data. If no new type appeared returns the original one.
    """
    max_id = master[id_column].max()
    list_new_values = [value for value in taxi_trips[value_column].values
                                  if value not in master[value_column].values]
    new_values_df = pd.DataFrame({
        value_column: range(max_id + 1, max_id + len(list_new_values) + 1),
        id_column: list_new_values
        })
    master.columns
    updated_master = pd.concat([master, new_values_df], ignore_index=True)
    
    return updated_master

test_payment_type_master = update_master(taxi_trips=taxi_trips_payment_type_only,
                                         master=payment_type_master,
                                         id_column='payment_type_id',
                                         value_column='payment_type'
                                         )

test_payment_type_master

test_company_type_master = update_master(taxi_trips=taxi_trips_company_only,
                                         master=company_master,
                                         id_column='company_id',
                                         value_column='company'
                                         )

test_company_type_master

#### update texi_trips with the most recent company_master and payment_type master codes

def update_taxi_trips_with_master_data(taxi_trips: pd.DataFrame,
                                   payment_type_master: pd.DataFrame,
                                   company_master: pd.DataFrame) -> pd.DataFrame:
    """ Removes the payment_type and the company columns, and adds their ids

    Args:
        taxi_trips (pd.DataFrame):
            DataFrame holding the daily taxi trips.
        payment_type_master (pd.Dataframe): Payment type master table.
        company_master (pd.Dataframe): Company master table.

    Returns:
        pd.DataFrame: The updated DataFrame having ids, not payment type and company names.
    """
    taxi_trips_id = taxi_trips.merge(payment_type_master, on='payment_type')
    taxi_trips_id = taxi_trips_id.merge(company_master, on='company')
    taxi_trips_id.drop(['payment_type', 'company'], axis=1, inplace=True)
    
    return taxi_trips_id

taxi_trips_id = update_taxi_trips_with_master_data(taxi_trips= taxi_trips,
                                   payment_type_master= payment_type_master,
                                   company_master= company_master)


taxi_trips_id.sample(5)

#### weather transformation function

def transform_weather_data(weather_data: json) -> pd.DataFrame:
    """ Make transformations on the daily weather api response.

    Args:
        weather_data (json):
            The daily weather data from the Open Metao API.
        
    Returns:
        pd.DataFrame:
            A DataFrame representation of the weather data.
    """
    weather_data_filtered = {
        'datetime': weather_data['hourly']['time'],
        'temperature': weather_data['hourly']['temperature_2m'],
        'wind_speed': weather_data['hourly']['wind_speed_10m'],
        'rain': weather_data['hourly']['rain'],
        'precipitation': weather_data['hourly']['precipitation']
        }
    weather_df = pd.DataFrame(weather_data_filtered)
    weather_df['datetime'] = pd.to_datetime(weather_df['datetime'])
    
    return weather_df

# test
url = 'https://archive-api.open-meteo.com/v1/era5'
# current date
current_datetime = datetime.now()
# get the data two months before as a formatted string
formatted_datetime = (current_datetime - relativedelta(months=2)).strftime('%Y-%m-%d')
params = {
    'latitude': 41.85,
    'longitude': -87.65,
    'start_date': formatted_datetime,
    'end_date': formatted_datetime,
    'hourly': 'temperature_2m,wind_speed_10m,rain,precipitation'
    }

response = requests.get(url, params=params)
weather_data = response.json()



weather_data_df = transform_weather_data(weather_data)

weather_data_df
